Rename prototyping to core.py, obliterate decorators
