+++ engarde/decorators.py
@@ -1,124 +1,14 @@
 # -*- coding: utf-8 -*-
-from __future__ import (unicode_literals, absolute_import, division)
+from engarde.core import CheckSet
 
-from functools import wraps
+none_missing = CheckSet().none_missing
+is_shape = CheckSet().is_shape
+unique_index = CheckSet().unique_index
+is_monotonic = CheckSet().is_monotonic
+within_set = CheckSet().within_set
+within_range = CheckSet().within_range
+within_n_std = CheckSet().within_n_std
+has_dtypes = CheckSet().has_dtypes
 
-import engarde.checks as ck
-
-def none_missing():
-    """Asserts that no missing values (NaN) are found"""
-    def decorate(func):
-        @wraps(func)
-        def wrapper(*args, **kwargs):
-            result = func(*args, **kwargs)
-            ck.none_missing(result)
-            return result
-        return wrapper
-    return decorate
-
-
-def is_shape(shape):
-    def decorate(func):
-        @wraps(func)
-        def wrapper(*args, **kwargs):
-            result = func(*args, **kwargs)
-            ck.is_shape(result, shape)
-            return result
-        return wrapper
-    return decorate
-
-
-def unique_index():
-    def decorate(func):
-        @wraps(func)
-        def wrapper(*args, **kwargs):
-            result = func(*args, **kwargs)
-            ck.unique_index(result)
-            return result
-        return wrapper
-    return decorate
-
-def is_monotonic(items=None, increasing=None, strict=False):
-    def decorate(func):
-        @wraps(func)
-        def wrapper(*args, **kwargs):
-            result = func(*args, **kwargs)
-            ck.is_monotonic(result, items=items, increasing=increasing,
-                            strict=strict)
-            return result
-        return wrapper
-    return decorate
-
-def within_set(items):
-    """
-    Check that DataFrame values are within set.
-
-    @within_set({'A': {1, 3}})
-    def f(df):
-        return df
-    """
-    def decorate(func):
-        @wraps(func)
-        def wrapper(*args, **kwargs):
-            result = func(*args, **kwargs)
-            ck.within_set(result, items)
-            return result
-        return wrapper
-    return decorate
-
-
-def within_range(items):
-    """
-    Check that a DataFrame's values are within a range.
-
-    Parameters
-    ==========
-
-    items : dict or array-like
-        dict maps columss to (lower, upper)
-        array-like checks the same (lower, upper) for each column
-
-    Raises
-    ======
-
-    ValueError
-    """
-    def decorate(func):
-        @wraps(func)
-        def wrapper(*args, **kwargs):
-            result = func(*args, **kwargs)
-            ck.SlicedChecks().within_range(result, items)
-            return result
-        return wrapper
-    return decorate
-
-
-def within_n_std(n=3):
-    """
-    Tests that all values are within 3 standard deviations
-    of their mean.
-    """
-    def decorate(func):
-        @wraps(func)
-        def wrapper(*args, **kwargs):
-            result = func(*args, **kwargs)
-            ck.within_n_std(result, n=n)
-            return result
-        return wrapper
-    return decorate
-
-def has_dtypes(items):
-    """
-    Tests that the dtypes are as specified in items.
-    """
-    def decorate(func):
-        @wraps(func)
-        def wrapper(*args, **kwargs):
-            result = func(*args, **kwargs)
-            ck.has_dtypes(result, items)
-            return result
-        return wrapper
-    return decorate
-
-__all__ = [is_monotonic, is_shape, none_missing, unique_index, within_n_std,
+__all__ = [none_missing, is_shape, none_missing, unique_index, within_n_std,
            within_range, within_set, has_dtypes]
